feat: add dashboard admin feature & and fix bug

+++ src/controllers/whitelist.controller.js
@@ -2,6 +2,7 @@ import WhitelistService from '../services/whitelist.service.js';
 import ResponseError from '../responses/error.response.js';
 import ResponseApi from '../responses/api.response.js';
 import * as whitelistValidate from '../validate/whitelist.validate.js';
+import {isValidObjectId} from 'mongoose';
 
 export default class WhitelistController {
   constructor() {
@@ -49,7 +50,18 @@ export default class WhitelistController {
         throw new ResponseError(error.message, 400);
       }
 
-      const whitelist = await this.service.deleteWhitelist(value.id);
+      // check if the whitelist exists and belongs to the user who is deleting it
+      const whitelist = await this.service.findByInovationId(value.id);
+      if (!whitelist) {
+        throw new ResponseError('Whitelist not found', 404);
+      }
+      if (isValidObjectId(req.user._id)) {
+        // check if the whitelist belongs to the user
+        if (whitelist.user_id.toString() !== req.user._id) {
+          throw new ResponseError('Unauthorized', 401);
+        }
+      }
+      await this.service.deleteWhitelist(value.id);
 
       return ResponseApi.noContent(res, whitelist);
     } catch (error) {
